2006-01-25  Martin Baulig  <martin@ximian.com>
	* backends/MonoThreadManager.cs: Started to clean this up; removed
	the `thread_hash'.

	* backends/Inferior.cs (Inferior.TID): Make this a `long', not `int'.
	* backends/Process.cs (Process.TID): Likewise.

	* backends/server/library.c
	(mono_debugger_server_attach): `tid' is now a `guint64 *'.
	(mono_debugger_server_peek_word): `word' is now a `guint64 *'.

	* backends/server/x86-arch.h (x86_arch_get_tid): Return a `guint64'.


svn path=/trunk/debugger/; revision=56023

+++ backends/SingleSteppingEngine.cs
@@ -1627,7 +1627,7 @@ namespace Mono.Debugger.Backends
 		public readonly string Name;
 		public readonly int ID;
 		public readonly int PID;
-		public readonly int TID;
+		public readonly long TID;
 
 		int stepping_over_breakpoint;
