+++ backends/server/x86-arch.h
@@ -36,7 +36,7 @@ x86_arch_child_stopped (ServerHandle *handle, int stopsig,
 static ServerCommandError
 x86_arch_get_registers (ServerHandle *handle);
 
-static guint32
+static guint64
 x86_arch_get_tid (ServerHandle *handle);
 
 #if defined(__i386__)
